src/scheduler/requirements.txt
pymongo
sqlalchemy
mysql-connector-python
kafka-python